Work in progress - moving to/from json to model level
