+++ examples/serialization.py
@@ -59,10 +59,11 @@ def run_example():
 
     # METHOD 1: Serialize with JSON 
     ### Step 4: serialize model for future use 
-    save_surrogate = surrogate_orig.parameters.to_json()
+    # save_surrogate = surrogate_orig.parameters.to_json()
+    save_surrogate = surrogate_orig.to_json()
 
     # DMDModel can be called directly with serialized result
-    surrogate_serial_1 = DMDModel.from_json(save_surrogate)
+    # surrogate_serial_1 = DMDModel.from_json(save_surrogate)
 
     # Serialized result can also be saved to a text file and uploaded later using the following code:
     txtFile = open("surrogate_model_save_json.txt", "w")
